improvement: chartcolor modify naming
