+++ src/components/Color/VisualColor.md
@@ -1,0 +1,9 @@
+### Visual Color
+
+```js
+import React, { useState } from 'react'
+import { VisualColors } from '@components/Color/VisualColor'
+const render = () => <VisualColors />
+
+render()
+```
